chore: move wordpressUtils to `serverUtils`
