+++ baker/pageOverrides.test.ts
@@ -1,8 +1,5 @@
-import {
-    FullPost,
-    WP_PostType,
-    extractFormattingOptions,
-} from "@ourworldindata/utils"
+import { FullPost, WP_PostType } from "@ourworldindata/utils"
+import { extractFormattingOptions } from "../serverUtils/wordpressUtils.js"
 import * as pageOverrides from "./pageOverrides.js"
 
 import { jest } from "@jest/globals"
